feat: jwt-16강 완료

+++ src/main/java/com/ll/sbbrestapi20250106/standard/util/Ut.java
@@ -28,7 +28,7 @@ public class Ut {
     }
 
     public static class jwt {
-        public static String toString(String secret, int expireSeconds, Map<String, Object> body) {
+        public static String toString(String secret, long expireSeconds, Map<String, Object> body) {
 
             Date issuedAt = new Date();
             Date expiration = new Date(issuedAt.getTime() + 1000L * expireSeconds);
